refactor(ui): разделён source management view layer
SourceManagementScreenView оставлен root navigation container
destination routing вынесен в SourceManagementScreenDestinationFactory
add feed, create folder, move source и shared source-management views разнесены по отдельным файлам без изменения screen flow
