+++ README.md
@@ -527,7 +527,7 @@
 - [x] `Reader Opening Contract Rename`: без legacy migration support переименовать `ReaderMode` / `defaultReaderMode` в контракт открытия статьи, который прямо описывает выбор между app-owned feed reader и `SFSafariViewController`; убрать оставшуюся browser/reader-mode терминологию, которая была уместна до удаления `WKWebView` и full-text reader extension point;
 - [x] `Unread Sort Contract Tightening`: сузить naming сортировки до unread-specific semantics на границе settings / screen state / query selection: убрать остаточную общую терминологию `sortMode` там, где она фактически означает настройку `Sort Unread Articles`, но не ломать repository-level сортировку, если она остаётся универсальным query parameter;
 - [x] `Source Management State Partition`: разрезать `SourceManagementScreenState` на отдельные state/reducer файлы для `Add Feed`, `Move Source` и `Create Folder`, оставив root state только координатором presentation destination и derived view state; сохранить текущий behavior, но сделать каждый flow тестируемым и review-friendly отдельно;
-- [ ] `Source Management View Split`: разделить крупный `SourceManagementScreenView` на root navigation container, destination factory и отдельные view-файлы для add/edit feed, move source и create/edit folder, чтобы UI-слой не смешивал navigation plumbing, bindings и конкретную разметку форм;
+- [x] `Source Management View Split`: разделить крупный `SourceManagementScreenView` на root navigation container, destination factory и отдельные view-файлы для add/edit feed, move source и create/edit folder, чтобы UI-слой не смешивал navigation plumbing, bindings и конкретную разметку форм;
 - [ ] `Reader Content Rendering Pipeline Split`: разнести `ArticleScreenContentRenderer` на доменные части: body block models, HTML block tokenizer / parser, inline text renderer, media URL resolver и final body assembly; renderer должен остаться thin orchestration layer между `ArticleScreenBodyPayloadNormalizer` и UI presentation blocks;
 - [ ] `Feed Parsing Module Split`: разделить `FeedParserService` на XML document builder, feed kind detection, RSS parser, Atom parser и parse pipeline diagnostics, чтобы будущие tests для parsing / normalizer не зависели от одного большого файла с несколькими уровнями ответственности;
 - [ ] `Refresh Pipeline Boundary Cleanup`: проверить `FeedRefreshService` после накопленных archive / retention / diagnostics изменений и выделить из него отдельные шаги conditional request, not-modified handling, fetched response reconciliation, diagnostics/log persistence и batch orchestration, не меняя публичный `FeedRefreshCoordinating` contract без необходимости;
